TIMPALA-146: Use multi-threaded scan nodes with mt_dop > 0
This patch add a new query option DISABLE_SCAN_NODE_MT, default
to false. If it is set true, the single-threaded scan node
(i.e. HdfsScanNodeMt or KuduScanNodeMt) will not be used even if
mt_dop > 0. So we can create more scanner threads to achieve
faster scanning with mt_dop > 0. Also, the type of scan node is
addied to the EXPLAIN output.

Tetting:
  - Added e2e tests to verify the new query option.

Change-Id: I40841eaaeee1756808aa5978cb58517ffd47c040