IMPALA-12312: Using correct executor group set info for planning
Prior to this patch, planner always selects the default group if there
is a default group in an impala cluster. When a client sets a
non-default request pool, planner still assumes the query run on the
default group and calculates the wrong number of nodes and instances.

This patch fixes it by including both default and non-default groups
in the update message sent from BE to FE, so planner can generate a
plan based on correct executor group set info. Besides, if no matched
executor group is found, planner falls back to using the default group
for planning, which is consistent with BE's behavior in
GetExecutorGroupsForQuery.

Tests:
  - Add new test cases to ClusterMembershipMgrUnitTest.
  - Add e2e test to verify the new behavior of planner.

Change-Id: Ia13fb40558441d4dcc0b3e7910d3746bb61e6b80